Enable RuntimeReporter's Page Source support for mobile applications
