+++ client/src/main/java/com/paypal/selion/internal/reports/model/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Model classes to assist in test reporting.
+ */
+package com.paypal.selion.internal.reports.model;
