+++ client/src/test/java/com/paypal/selion/android/sample/AppiumAndroidBrowserSauceCloudTest.java
@@ -23,7 +23,7 @@ import org.testng.annotations.Test;
 
 import com.paypal.selion.annotations.MobileTest;
 import com.paypal.selion.platform.grid.Grid;
-import com.paypal.selion.reports.runtime.MobileReporter;
+import com.paypal.selion.reports.runtime.SeLionReporter;
 
 /*
  * DEVNOTE Tests in this class exist primarily for demonstration purposes and as a basic sanity checks.
@@ -43,7 +43,7 @@ public class AppiumAndroidBrowserSauceCloudTest {
         element.sendKeys("Cheese!");
         // Now submit the form. WebDriver will find the form for us from the element
         element.submit();
-        MobileReporter.log("cheese!", true);
+        SeLionReporter.log("cheese!", true);
 
     }
